+++ src/ArduinoJson/Deserialization/deserialize.hpp
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <ArduinoJson/Deserialization/DeserializationError.hpp>
+#include <ArduinoJson/Deserialization/Filter.hpp>
 #include <ArduinoJson/Deserialization/NestingLimit.hpp>
 #include <ArduinoJson/Deserialization/Reader.hpp>
 #include <ArduinoJson/StringStorage/StringStorage.hpp>
@@ -19,47 +20,52 @@ TDeserializer<TReader, TWriter> makeDeserializer(MemoryPool &pool,
   return TDeserializer<TReader, TWriter>(pool, reader, writer, nestingLimit);
 }
 
-// deserialize(JsonDocument&, const std::string&);
-// deserialize(JsonDocument&, const String&);
-// deserialize(JsonDocument&, char*);
-// deserialize(JsonDocument&, const char*);
-// deserialize(JsonDocument&, const __FlashStringHelper*);
-template <template <typename, typename> class TDeserializer, typename TString>
+// deserialize(JsonDocument&, const std::string&, NestingLimit, Filter);
+// deserialize(JsonDocument&, const String&, NestingLimit, Filter);
+// deserialize(JsonDocument&, char*, NestingLimit, Filter);
+// deserialize(JsonDocument&, const char*, NestingLimit, Filter);
+// deserialize(JsonDocument&, const __FlashStringHelper*, NestingLimit, Filter);
+template <template <typename, typename> class TDeserializer, typename TString,
+          typename TFilter>
 typename enable_if<!is_array<TString>::value, DeserializationError>::type
-deserialize(JsonDocument &doc, const TString &input,
-            NestingLimit nestingLimit) {
+deserialize(JsonDocument &doc, const TString &input, NestingLimit nestingLimit,
+            TFilter filter) {
   Reader<TString> reader(input);
   doc.clear();
   return makeDeserializer<TDeserializer>(
              doc.memoryPool(), reader,
              makeStringStorage(doc.memoryPool(), input), nestingLimit.value)
-      .parse(doc.data());
+      .parse(doc.data(), filter);
 }
 //
-// deserialize(JsonDocument&, char*, size_t);
-// deserialize(JsonDocument&, const char*, size_t);
-// deserialize(JsonDocument&, const __FlashStringHelper*, size_t);
-template <template <typename, typename> class TDeserializer, typename TChar>
+// deserialize(JsonDocument&, char*, size_t, NestingLimit, Filter);
+// deserialize(JsonDocument&, const char*, size_t, NestingLimit, Filter);
+// deserialize(JsonDocument&, const __FlashStringHelper*, size_t, NL, Filter);
+template <template <typename, typename> class TDeserializer, typename TChar,
+          typename TFilter>
 DeserializationError deserialize(JsonDocument &doc, TChar *input,
-                                 size_t inputSize, NestingLimit nestingLimit) {
+                                 size_t inputSize, NestingLimit nestingLimit,
+                                 TFilter filter) {
   BoundedReader<TChar *> reader(input, inputSize);
   doc.clear();
   return makeDeserializer<TDeserializer>(
              doc.memoryPool(), reader,
              makeStringStorage(doc.memoryPool(), input), nestingLimit.value)
-      .parse(doc.data());
+      .parse(doc.data(), filter);
 }
 //
-// deserialize(JsonDocument&, std::istream&);
-// deserialize(JsonDocument&, Stream&);
-template <template <typename, typename> class TDeserializer, typename TStream>
+// deserialize(JsonDocument&, std::istream&, NestingLimit, Filter);
+// deserialize(JsonDocument&, Stream&, NestingLimit, Filter);
+template <template <typename, typename> class TDeserializer, typename TStream,
+          typename TFilter>
 DeserializationError deserialize(JsonDocument &doc, TStream &input,
-                                 NestingLimit nestingLimit) {
+                                 NestingLimit nestingLimit, TFilter filter) {
   Reader<TStream> reader(input);
   doc.clear();
   return makeDeserializer<TDeserializer>(
              doc.memoryPool(), reader,
              makeStringStorage(doc.memoryPool(), input), nestingLimit.value)
-      .parse(doc.data());
+      .parse(doc.data(), filter);
 }
+
 }  // namespace ARDUINOJSON_NAMESPACE